Added support for user defined plugins handling from within the config system. Should fix ticket 205.
Three the new entries to the config system :

server.filtersRoot : specify where CherryPy should look for finding filters modules, it's a list of strings. Nothing is prefixed to those paths
server.inputFiltersDict : a dictionnary that defines which filters should act on the output. The keys are Python module (without the .py extension), the value is the classname within the module
server.outputFiltersDict : a dictionnary that defines which filters should act on the output. The keys are Python module (without the .py extension), the value is the classname within the module

For an example, look at the documentation